install TypeScript and convert index.js to index.ts
