+++ package.json
@@ -10,7 +10,7 @@
     "start": "echo The app is starting! It will automatically open in your browser when ready && env-cmd -f ./.env parcel src/index.html --open",
     "dev": "nodemon --ext js --exec \"env-cmd -f ./.env.local parcel src/index.html --open\" --ignore \"dist\"",
     "test": "jest test --runInBand",
-    "start:server": "(cd server && yarn start-dev)"
+    "start:server": "(cd server && yarn dev)"
   },
   "devDependencies": {
     "@babel/core": "~7.14.0",
