+++ server/README.md
@@ -1,10 +1,9 @@
 # Data Collector and API server
 
-Quick Start
-===========
+# Quick Start
 
 To run this project locally:
 
 1. Run the MongoDB server `docker-compose up -d`
 2. Install dependencies: `yarn install`
-3. Run the API server: `yarn start-dev`
+3. Run the API server: `yarn dev`
