+++ server/package.json
@@ -1,11 +1,11 @@
 {
   "name": "near-transactions-accounting-report",
   "version": "0.1.0",
-  "main": "src/index.js",
+  "main": "src/index",
   "type": "module",
   "scripts": {
-    "start": "node src/index.js",
-    "start-dev": "nodemon -r dotenv/config src/index.js dotenv_config_path=.env.development.local",
+    "start": "node src/index",
+    "dev": "node -r dotenv/config --loader ts-node/esm src/index.ts dotenv_config_path=.env.development.local",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "dependencies": {
@@ -20,6 +20,7 @@
   },
   "devDependencies": {
     "nodemon": "^2.0.15",
-    "prettier": "^2.5.1"
+    "prettier": "^2.5.1",
+    "typescript": "^4.8.2"
   }
 }
